Revert "Added support for connecting to TLS enabled memcached (new PR)"
