+++ folsom/src/main/java/com/spotify/folsom/MemcacheClientBuilder.java
@@ -37,7 +37,6 @@ import com.spotify.folsom.client.NoopMetrics;
 import com.spotify.folsom.client.NoopTracer;
 import com.spotify.folsom.client.ascii.DefaultAsciiMemcacheClient;
 import com.spotify.folsom.client.binary.DefaultBinaryMemcacheClient;
-import com.spotify.folsom.client.tls.SSLEngineFactory;
 import com.spotify.folsom.guava.HostAndPort;
 import com.spotify.folsom.ketama.AddressAndClient;
 import com.spotify.folsom.ketama.KetamaMemcacheClient;
@@ -121,8 +120,6 @@ public class MemcacheClientBuilder<V> {
 
   private final List<UsernamePasswordPair> passwords = new ArrayList<>();
   private boolean skipAuth = false;
-
-  private SSLEngineFactory sslEngineFactory = null;
 
   /**
    * Create a client builder for byte array values.
@@ -559,11 +556,6 @@ public class MemcacheClientBuilder<V> {
     return this;
   }
 
-  public MemcacheClientBuilder<V> withSSLEngineFactory(final SSLEngineFactory sslEngineFactory) {
-    this.sslEngineFactory = sslEngineFactory;
-    return this;
-  }
-
   /**
    * Create a client that uses the binary memcache protocol.
    *
@@ -716,7 +708,6 @@ public class MemcacheClientBuilder<V> {
         metrics,
         maxSetLength,
         eventLoopGroup,
-        channelClass,
-        sslEngineFactory);
+        channelClass);
   }
 }
