--- folsom/src/main/java/com/spotify/folsom/client/tls/SSLEngineFactory.java
@@ -1,7 +1,0 @@
-package com.spotify.folsom.client.tls;
-
-import javax.net.ssl.SSLEngine;
-
-public interface SSLEngineFactory {
-  SSLEngine createSSLEngine(String hostname, int port);
-}
