+++ folsom/src/main/java/com/spotify/folsom/reconnect/ReconnectingClient.java
@@ -26,7 +26,6 @@ import com.spotify.folsom.authenticate.Authenticator;
 import com.spotify.folsom.client.DefaultRawMemcacheClient;
 import com.spotify.folsom.client.NotConnectedClient;
 import com.spotify.folsom.client.Request;
-import com.spotify.folsom.client.tls.SSLEngineFactory;
 import com.spotify.folsom.guava.HostAndPort;
 import com.spotify.folsom.ketama.AddressAndClient;
 import io.netty.channel.Channel;
@@ -75,8 +74,7 @@ public class ReconnectingClient extends AbstractRawMemcacheClient {
       final Metrics metrics,
       final int maxSetLength,
       final EventLoopGroup eventLoopGroup,
-      final Class<? extends Channel> channelClass,
-      final SSLEngineFactory sslEngineFactory) {
+      final Class<? extends Channel> channelClass) {
     this(
         backoffFunction,
         scheduledExecutorService,
@@ -92,8 +90,7 @@ public class ReconnectingClient extends AbstractRawMemcacheClient {
                 metrics,
                 maxSetLength,
                 eventLoopGroup,
-                channelClass,
-                sslEngineFactory),
+                channelClass),
         authenticator,
         address);
   }
