+++ folsom/src/test/java/com/spotify/folsom/MemcachedServer.java
@@ -16,8 +16,6 @@
 package com.spotify.folsom;
 
 import com.google.common.base.Suppliers;
-import com.spotify.folsom.client.tls.DefaultSSLEngineFactory;
-import java.security.NoSuchAlgorithmException;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
@@ -40,64 +38,34 @@ public class MemcachedServer {
       Suppliers.memoize(MemcachedServer::new)::get;
 
   public static int DEFAULT_PORT = 11211;
-
-  private static final String MEMCACHED_VERSION = "1.6.21";
   private final String username;
   private final String password;
-  private final boolean useTLS;
 
   public MemcachedServer() {
     this(null, null);
   }
 
   public MemcachedServer(String username, String password) {
-    this(username, password, DEFAULT_PORT, AuthenticationMode.NONE, false);
+    this(username, password, DEFAULT_PORT, AuthenticationMode.NONE);
   }
 
   public MemcachedServer(String username, String password, AuthenticationMode authenticationMode) {
-    this(username, password, DEFAULT_PORT, authenticationMode, false);
-  }
-
-  public MemcachedServer(boolean useTLS) {
-    this(null, null, DEFAULT_PORT, AuthenticationMode.NONE, useTLS);
+    this(username, password, DEFAULT_PORT, authenticationMode);
   }
 
   public MemcachedServer(
-      String username,
-      String password,
-      int port,
-      AuthenticationMode authenticationMode,
-      boolean useTLS) {
+      String username, String password, int port, AuthenticationMode authenticationMode) {
     this.username = username;
     this.password = password;
-    this.useTLS = useTLS;
-
-    if (useTLS) {
-      if (authenticationMode != AuthenticationMode.NONE) {
-        throw new RuntimeException(
-            "Authentication is not currently supported with the TLS-enabled container");
-      }
-
-      container = setupTLSContainer();
-    } else {
-      container = setupContainer(username, password, authenticationMode);
-    }
-
+    container = new FixedHostPortGenericContainer("bitnami/memcached:1.6.21");
     if (port != DEFAULT_PORT) {
       container.withFixedExposedPort(DEFAULT_PORT, port);
     } else {
       container.withExposedPorts(DEFAULT_PORT);
     }
-
-    start(authenticationMode);
-  }
-
-  private FixedHostPortGenericContainer setupContainer(
-      String username, String password, AuthenticationMode authenticationMode) {
-    final FixedHostPortGenericContainer container =
-        new FixedHostPortGenericContainer("bitnami/memcached:" + MEMCACHED_VERSION);
-
     switch (authenticationMode) {
+      case NONE:
+        break;
       case SASL:
         if (username != null && password != null) {
           container.withEnv("MEMCACHED_USERNAME", username);
@@ -110,30 +78,7 @@ public class MemcachedServer {
         break;
     }
 
-    return container;
-  }
-
-  private FixedHostPortGenericContainer setupTLSContainer() {
-    final FixedHostPortGenericContainer container =
-        new FixedHostPortGenericContainer("memcached:" + MEMCACHED_VERSION);
-
-    container.withClasspathResourceMapping(
-        "/pki/test.pem", "/test-certs/test.pem", BindMode.READ_ONLY);
-    container.withClasspathResourceMapping(
-        "/pki/test.key", "/test-certs/test.key", BindMode.READ_ONLY);
-
-    container.withCommand(
-        "memcached",
-        "--enable-ssl",
-        "-o",
-        "ssl_verify_mode=3",
-        "-o",
-        "ssl_chain_cert=/test-certs/test.pem",
-        "-o",
-        "ssl_key=/test-certs/test.key",
-        "-o",
-        "ssl_ca_cert=/test-certs/test.pem");
-    return container;
+    start(authenticationMode);
   }
 
   public void stop() {
@@ -157,15 +102,6 @@ public class MemcachedServer {
       if (username != null && password != null) {
         builder.withUsernamePassword(username, password);
       }
-
-      if (useTLS) {
-        try {
-          builder.withSSLEngineFactory(new DefaultSSLEngineFactory(false));
-        } catch (NoSuchAlgorithmException e) {
-          throw new RuntimeException(e);
-        }
-      }
-
       switch (authenticationMode) {
         case NONE:
         case SASL:
