--- folsom/src/test/resources/pki/README
@@ -1,5 +1,0 @@
-These test certificates were generated as follows:
-
-openssl req -x509 -newkey rsa:4096 -keyout test.key -out test.pem -sha256 -days 3650 -nodes -subj '/CN=localhost'
-openssl pkcs12 -export -inkey test.key -in test.pem -out test.p12 -password pass:changeit
-
