--- folsom/src/test/resources/pki/test.key
@@ -1,52 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIJQgIBADANBgkqhkiG9w0BAQEFAASCCSwwggkoAgEAAoICAQDCkosiWmgj0h1T
-FbyX6tNDbdiiePM76+SpjwnJvdIHxosqgzDaetEGcXHAiD+5jhUlGzSjJv8GzcXs
-vRXSaCjnhsOQya1xJxQqPxWJtFOAgTgJfWvVRJWh5t/RQyZb6/mCvGkV4xmA/spF
-ahGwbeLvQhWmsmYpUdXWZOOzfEONkz8DD4qY5NMOyF1XYR6vJTjwxYOjJqSXFMXb
-7CdP/shiMFW1o5vt6FIDIGempIqX11mH7xa82gXn0AaJf/XbDFLkhOaqYIE01np/
-ZCQB+fhpTo/K+4iJuH2WEsF+h6rz/4TukrT+tZNZJDeiBFX1SiT2jUIrXc6m/8MA
-op1R8jWmIi1tKQe+EGFQg1RJ7tdrxuTzFF6ZIPuOO+O3uuNDiaF4nD3Fb4/ihdwf
-MCNEmneGsB56vvcovMk73ygY8najdKgJ9Q8Zzt63LuDkT2xtbb1IEgWQCD2tIx7p
-rB0FPK1kAtVqdYpkr1K8z9jHHLXVpYg24U/nRdLDjmxzsJy0UsQjA08baT5Z+8lj
-H7pZiCr/Oua+Xqyf04+YLl/BDzRVJxU9/PJGm9J7ToK2W2hbOb5ZyGi0A/gYie+l
-Mt/dAaK0NaOvT+LPU9h4QBSDT7+EKTovAVQV9oaCPMajMYEQM97SlYIRTgfdxjz/
-XWBjK7O+5xHYDB3tQ2yDsBe6J2JTBQIDAQABAoICAFURVu1dE3zdx15k+YB99YHc
-i8N1F/sRxnZviDsA18v4fS3ID9rlqW+kt7QSnbHVsd74Rwox6XwjCS7Y+Kp2SbP4
-EpbB5ie7izBxRkxfX8amOEbF5BhpFPalByPptOKph/wlvdgI40SnoO26UyOM15P9
-k5O/GbVlBxXmJDt7z9kdEIaZC+KO4MvsxAqI9q5imsOBx2zUX0+CkIL3e7SY0ylC
-sqZocqsQUJL3XvnXpJSAXa91074boEtagxEotefgOnmYtXS2HqdoQkRiFvAaRwZb
-h7iaQnbcB2ROrPRSAmcNRyQ7VhCqeFKX6A7Y0HAey4xT9CMbr1cKKUqkzh1exR8p
-1xAKoCZ2BxycdF5PSDc4aYGMC4vgjCUR4ucKRLjY9+uVsrhNx2C+VwLKLWkABEd1
-NBV3oQ0OiS3vgUBArMBzSnng2xMr4sP4Cjnv/spuNBe18QDIJOKMze2hyUA6JHT/
-A6crY2ypCspp6Ik2APjTMbgRW4E+g9U7+0rgAgr0EFsFZUJ4R6I4oPaNjh8mpVwd
-KBx71JiAjlmNWqfWzkThMzbJfeEbmcONJf0/QTyQDPtazw8WIGzq1dL1lcLLibs0
-T0FC1woHVnuPvbD/deBtYoWBseItYAWNqJzDJi/O10eAc7bR4MT4aq3t3CVaxZfC
-Ea6lPlQpi/XP0S4/k/MBAoIBAQD8rPeL/9VdUcHDZXO86aGUEEnzXn6EhDDbA5k4
-0mBHmk17DpJ7crB8I/fyyjiB9YVBx5e+27s1S6EnvIdDO0rUV4mZLvtiv630bp3D
-R1TP4x01jK2CarLI70tBTnfxDH3E5PZks8bN2u9TPipwB93/LV3HIRh6G6P1iAmw
-O3KCw0qWtyPPr5pagsitt+ulXunZY6hrnjpV1hRJkVugJ0/906WKb1GFOdfys8tx
-V5z/hSuJTP40ijyREkXyXOkqVKQk5xNpmAL8ZFG5JELwgjXSdezNTWwA5i0q26pD
-OP9X8nWvoLY22Kqp8TX8Rb9+PMjSvxmW5bfGeVLYHZ2HytvBAoIBAQDFIeFEk+TR
-/uzA3wrpYpFun23n+snjIQQtGa9MhhADeIPzaQXmwoO/FTIn6YZ8S3A7dikO7wtK
-DhIli4nyqQ+GGe/bvw+b+98GBF0Ym0KL1epWBQQSCT1mRSWkr37akLoJ70pOOxo+
-pErZanjOgQBpv1Kw66RP5DobhxXs29j7sLF9ai3aq9f985485ZdhOYGisjNV4412
-ccLnB3D0hbN+5sR/OnLYjB6aDxl4QzBYWsMEQMfEVjxvfYQIGgW+OuRCWtVv1gRC
-XR5HUE76g1zEC3w/qghAdbclkn/RIzz+xqW3a0Qmgr0Os2x0rcndRgySDeyUEhKp
-nMXvfh5L2xhFAoIBAFNTBt9YIph/mZJCJoSp1uro5DopczdoEGRpL2IZnj5+mAZ4
-q72h+Kk3g1DBdxkESkmC9HuwInBU3HQqK8D0EJ0tsOafI69Q3qC4ybXYFBPqJXu/
-RIi5fvPcVcjXg54uLFt91fMnhevkwv+EhKIlNgQshbxhIZ1C+DLEBc3kDMMqe7Jv
-+pNGqXQnpN4ExOToA934i1XR/BLKYi5QjRKnZC2kWfbo9s0kYh5bRD/AULnCxLSm
-ez4ASDDfAcoG8a1P9EFnInOz+WgZ/Qk8+AYwKmeZE9owKemx/jsf7Wn1pd7uyfh8
-2xoDIWShctgaeCe9C8zT0DB+2LfO0o8KVSSutQECggEAQMRs2rcsisISzxt43k0A
-MzQ2S+1dvz3tvVOfAKlbQYiq8aIjyjlGR9WS8QIMqXGvohmkS7/GGcKdu7Ao1o0t
-CIYlBDG10y3hjHyKibcZGhBiOXjUaYiXn65AO+dc6jp6pSD1bNaGPOaFoQEWR+Ki
-XBv78xy4k3cMkFbFoVhp5eebqPTls96ZzFnqN1/HaK4YJXge3a0xoSSnQHh1aCE1
-ZBA1pwdxDCydMUicuaJ5k83eHNubxqn+mTLH2lGSaXm33QUy8teB9rvZYtzD1hKq
-u856OACJTYRfc/y5+eB1/c8OS0D21yBFNTtF+t/OXuDQ6HuiqtN1Rjy324O4OHv3
-cQKCAQEA/GrY7PLQlkWLsNj1xwzi5Vb4H4DaHmYGtTGi8LIIcYFQw0HN7khb7WNP
-3FKqCR8Zt+mZ60IZDShggiBtuexc7ghsgF29MvhX2Kx9BdBFuGS3C/Sy+wwxd7Ln
-8RvCcpII151G5oDL/sXTEfMa7+NIN2CAqPgTDt28vbQuyIXbI0xifbhZuvv76be6
-0UUBfAqxJzo7jo/EfC7wEOOOyw4vgHcOwHrbN6rmsTd4/OReRGEhMIawcou+pziY
-r+ZS1xXUZxI5eUVmFK47BSgzaZLAOvbWQ+ZFVScxQOpiEI7wXbp7vDHvPzNJ2pYX
-G6gcbkorel8YtBdPnaweR3yhIwhcyA==
------END PRIVATE KEY-----
