--- folsom/src/test/resources/pki/test.pem
@@ -1,29 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIFCTCCAvGgAwIBAgIUW7SurLbUDUtQfw3EdyDv8HGmcZMwDQYJKoZIhvcNAQEL
-BQAwFDESMBAGA1UEAwwJbG9jYWxob3N0MB4XDTIzMDkwNzA3MTczMVoXDTMzMDkw
-NDA3MTczMVowFDESMBAGA1UEAwwJbG9jYWxob3N0MIICIjANBgkqhkiG9w0BAQEF
-AAOCAg8AMIICCgKCAgEAwpKLIlpoI9IdUxW8l+rTQ23YonjzO+vkqY8Jyb3SB8aL
-KoMw2nrRBnFxwIg/uY4VJRs0oyb/Bs3F7L0V0mgo54bDkMmtcScUKj8VibRTgIE4
-CX1r1USVoebf0UMmW+v5grxpFeMZgP7KRWoRsG3i70IVprJmKVHV1mTjs3xDjZM/
-Aw+KmOTTDshdV2EeryU48MWDoyaklxTF2+wnT/7IYjBVtaOb7ehSAyBnpqSKl9dZ
-h+8WvNoF59AGiX/12wxS5ITmqmCBNNZ6f2QkAfn4aU6PyvuIibh9lhLBfoeq8/+E
-7pK0/rWTWSQ3ogRV9Uok9o1CK13Opv/DAKKdUfI1piItbSkHvhBhUINUSe7Xa8bk
-8xRemSD7jjvjt7rjQ4mheJw9xW+P4oXcHzAjRJp3hrAeer73KLzJO98oGPJ2o3So
-CfUPGc7ety7g5E9sbW29SBIFkAg9rSMe6awdBTytZALVanWKZK9SvM/Yxxy11aWI
-NuFP50XSw45sc7CctFLEIwNPG2k+WfvJYx+6WYgq/zrmvl6sn9OPmC5fwQ80VScV
-PfzyRpvSe06CtltoWzm+WchotAP4GInvpTLf3QGitDWjr0/iz1PYeEAUg0+/hCk6
-LwFUFfaGgjzGozGBEDPe0pWCEU4H3cY8/11gYyuzvucR2Awd7UNsg7AXuidiUwUC
-AwEAAaNTMFEwHQYDVR0OBBYEFAd3Z5PTSftej1gL82mVQQuOVOYEMB8GA1UdIwQY
-MBaAFAd3Z5PTSftej1gL82mVQQuOVOYEMA8GA1UdEwEB/wQFMAMBAf8wDQYJKoZI
-hvcNAQELBQADggIBAJbVyzamDYqiIOSeHe19e8lWyLu7Zb50k7yyw406dbBaPRZf
-lBjyh2XWNdcVc/wjxdA9EDlIkIzNlXzwlyDa5VqPwzHdkel6dc6dqVQHzMWuoDzL
-HY1K7kMaRP3It0cDlJszsMI+kri+9wWEGkILJiNsKUJcjCFAQUxTe6vQWyLu63VY
-aY/9HgqkFqI5rWmLCYlXF5lCdDsMPmVGswBLQarJFrm6j4CxlwMArKfqceDvsJ17
-qPtDi/r6ItmTRv0xRAh1fJG0tRtyvx1mQ8t69GM1BFDrXYIlvhTXFDPE/DkypClo
-toQWKX6PIyQBhWNQL6JGrEGyeY8ljc/atRYWUugxQSuZ5EWdcwHZas4Z0eFYjPLa
-K0PYlFA3Y2bjzvMZVDNPggSDYB/RgfZQzGBNTqH4LJ8yqQhBMTkFlkHypSpAyru3
-tAMDQC5dLGeR1HNJcJxr9gU3PxmdRMSTRQxf9LtpznXnihZYmEjEnZ0ULKAN+8sq
-eaREI6B3/3aTKkoBUV028GxlHjwmAtp9zhwV6UCg2z8SItCFvcW95MNGu312MMy1
-fTBv3wDgaGqyy9ZddrANQGbXleJcrEJDaIHWdsJ0q077BuWN/c8EnuyqhyOWklEO
-ly2SJltjIA/P30BSpEyvb/MweSQ9LhNORfPrhGJotDmDHTw8mqbfsdUaVirq
------END CERTIFICATE-----
